+++ src/Utility.ts
@@ -2,12 +2,11 @@
 
 import * as child_process from "child_process";
 import * as fse from "fs-extra";
-import * as net from "net";
 import * as os from "os";
 import * as path from "path";
 import * as readline from "readline";
 import * as vscode from "vscode";
-import { Session, TelemetryWrapper } from "vscode-extension-telemetry-wrapper";
+import { instrumentOperationStep, sendInfo } from "vscode-extension-telemetry-wrapper";
 import * as xml2js from "xml2js";
 import * as Constants from "./Constants";
 import { DialogMessage } from "./DialogMessage";
@@ -44,15 +43,13 @@ export namespace Utility {
         }
         vscode.window.showTextDocument(vscode.Uri.file(file), { preview: false });
     }
-    export function trackTelemetryStep(step: string): void {
-        const session: Session = TelemetryWrapper.currentSession();
-        if (session && session.extraProperties) { session.extraProperties.finishedSteps.push(step); }
-        TelemetryWrapper.info(step);
+
+    export function trackTelemetryStep(operationId: string, step: string, callback: (...args: any[]) => any): any {
+        return instrumentOperationStep(operationId, step, callback)();
     }
 
-    export function initTelemetrySteps(): void {
-        const session: Session = TelemetryWrapper.currentSession();
-        if (session && session.extraProperties) { session.extraProperties.finishedSteps = []; }
+    export function infoTelemetryStep(operationId: string, step: string): void {
+        sendInfo(operationId, { finishedStep: step });
     }
 
     export function disableAutoRestart(): void {