+++ test/Tomcat.test.ts
@@ -5,14 +5,14 @@ import { DialogMessage } from '../src/DialogMessage';
 import { TomcatController } from "../src/Tomcat/TomcatController";
 import { TomcatModel } from "../src/Tomcat/TomcatModel";
 import { TomcatServer } from "../src/Tomcat/TomcatServer";
-import { Utility } from "../src/Utility";
+import { createUuid } from "vscode-extension-telemetry-wrapper";
 
 suite('Error input', () => {
   const serverInfo: TomcatServer = undefined;
   const tomcatModel: TomcatController = new TomcatController(new TomcatModel(''), undefined);
   test('stopServer', async () => {
     try {
-      await tomcatModel.stopOrRestartServer(serverInfo);
+      await tomcatModel.stopOrRestartServer(createUuid(), serverInfo);
       assert.fail('Resolve', 'Reject');
     } catch (error) {
       assert.equal(error.toString(), `Error: ${DialogMessage.noServer}`);
@@ -20,7 +20,7 @@ suite('Error input', () => {
   });
   test('runOnServer', async () => {
     try {
-      await tomcatModel.runOrDebugOnServer(undefined);
+      await tomcatModel.runOrDebugOnServer(createUuid(), undefined);
     } catch (error) {
       assert.equal(error.toString(), `Error: ${DialogMessage.noServer}`);
     }