+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 All notable changes to the "vscode-tomcat" extension will be documented in this file.
 
+# 0.11.2
+### Changed
+* Update the `vscode-extension-telemetry-wrapper` to `0.8.0`.
+
 # 0.11.1
 ### Fixed
-Account for case when looking at context XML [#218](https://github.com/adashen/vscode-tomcat/issues/#218)
+* Account for case when looking at context XML [#218](https://github.com/adashen/vscode-tomcat/issues/#218)
 
 # 0.11.0
 ### Added